// frontend-vite-app/src/components/view/personal_card/types/checkboxTypes.tsx
export interface CheckboxState {
    fullTime: boolean;
    partTime25: boolean;
    partTime50: boolean;
    partTime75: boolean;
    coWorker: boolean;
    externalCoWorker: boolean;
    temporaryWorker: boolean;
    postalPaidTeacher: boolean;
    scientificCandidate: boolean;
    scientificDoctor: boolean;
    academician: boolean;
    chernobyl1: boolean;
    chernobyl2: boolean;
    chernobyl3: boolean;
    chernobyl4: boolean;
    disability1: boolean;
    disability2: boolean;
    disability3: boolean;
}